Update TreasureFactory.cpp
This removes the Cases that append the Missile and Magic Casters with Elemental type/dmg/ui/name

As of recent changes in both content, the jsons provide this information already thus making the code irrelevant.
For this to work, the server needs to have the new TOD Elemental Missile and Caster jsons and the loot profile pointing to them.


Only bug I have found is that Casters such as the Baton are not being appended with Material in name/icon and palette.